Rename `MessageFactory` to `Given`
