+++ client/src/test/java/io/spine/examples/todolist/client/CommandLineTodoClientTest.java
@@ -43,8 +43,8 @@ import java.util.concurrent.TimeUnit;
 
 import static io.spine.base.Identifier.newUuid;
 import static io.spine.client.ConnectionConstants.DEFAULT_CLIENT_SERVICE_PORT;
-import static io.spine.examples.todolist.testdata.MessageFactory.newDescription;
 import static io.spine.examples.todolist.client.CommandLineTodoClient.HOST;
+import static io.spine.examples.todolist.testdata.Given.newDescription;
 import static io.spine.examples.todolist.testdata.TestLabelCommandFactory.LABEL_TITLE;
 import static io.spine.examples.todolist.testdata.TestTaskCommandFactory.DESCRIPTION;
