+++ testutil-api/src/main/java/io/spine/examples/todolist/testdata/TestTaskCommandFactory.java
@@ -40,7 +40,7 @@ import io.spine.examples.todolist.c.commands.UpdateTaskDueDate;
 import io.spine.examples.todolist.c.commands.UpdateTaskPriority;
 
 import static io.spine.base.Identifier.newUuid;
-import static io.spine.examples.todolist.testdata.MessageFactory.newDescription;
+import static io.spine.examples.todolist.testdata.Given.newDescription;
 import static io.spine.time.Time.getCurrentTime;
 
 /**
