+++ testutil-api/src/main/java/io/spine/examples/todolist/testdata/TestTaskEventFactory.java
@@ -43,7 +43,7 @@ import io.spine.examples.todolist.c.events.TaskPriorityUpdated;
 import io.spine.examples.todolist.c.events.TaskReopened;
 
 import static io.spine.base.Identifier.newUuid;
-import static io.spine.examples.todolist.testdata.MessageFactory.newDescription;
+import static io.spine.examples.todolist.testdata.Given.newDescription;
 import static io.spine.time.Time.getCurrentTime;
 
 /**
